pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "dev-health-ops"
dynamic = ["version"]
description = "Open Source Development Team Operation Analytics"
readme = "README.md"
authors = [{ name = "Chris Geo", email = "chris@example.com" }]
license = { text = "MIT" }
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
keywords = ["metrics", "git", "analytics", "developer-experience", "grafana"]
requires-python = ">=3.10"
dependencies = [
  "tqdm",
  "sqlalchemy",
  "clickhouse-connect",
  "fastapi",
  "httpx",
  "psycopg2-binary",
  "asyncpg",
  "aiosqlite",
  "python-gitlab",
  "GitPython",
  "PyGithub>=2.1.0,<3.0.0",
  "requests",
  "pydantic",
  "PyYAML",
  "typing_extensions",
  "jira",
  "alembic",
  "motor",
  "uvicorn",
  "radon",
  "openai",
  "anthropic",
  "strawberry-graphql[fastapi]",
  "celery[redis]>=5.3.0",
  "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
  "mypy",
  "black",
  "flake8",
  "isort",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
]

[project.urls]
Repository = "https://github.com/chrisgeo/dev-health-ops"
Issues = "https://github.com/chrisgeo/dev-health-ops/issues"

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "dirty-tag"

[tool.setuptools]
package-dir = { "" = "src" }
packages = { find = { where = ["src"], exclude = ["tests*", "docs*", "examples*", "data*", "grafana*", "docker*", "dist*", "build*"] } }
py-modules = ["cli", "investment_taxonomy", "storage", "utils"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["**/*.sql", "**/*.yaml", "**/*.yml", "**/*.json"]

[tool.poetry]
name = "dev-health-ops"
version = "0.1.0"
description = "Open Source Development Team Operation Analytics"
authors = ["Chris Geo <chris@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "dev_health_ops", from = "src" },
]
include = [
    "src/dev_health_ops/config/**/*",
    "src/dev_health_ops/alembic.ini",
    "src/dev_health_ops/alembic/**/*"
]

[tool.poetry.dependencies]
python = "^3.10"
tqdm = "*"
sqlalchemy = "*"
clickhouse-connect = "*"
fastapi = "*"
httpx = "*"
psycopg2-binary = "*"
asyncpg = "*"
aiosqlite = "*"
python-gitlab = "*"
GitPython = "*"
PyGithub = ">=2.1.0,<3.0.0"
requests = "*"
pydantic = "*"
PyYAML = "*"
typing_extensions = "*"
jira = "*"
alembic = "*"
motor = "*"
uvicorn = "*"
radon = "*"
openai = "*"
anthropic = "*"
strawberry-graphql = {extras = ["fastapi"], version = "*"}
celery = {extras = ["redis"], version = ">=5.3.0"}
redis = ">=5.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "*"
black = "*"
flake8 = "*"
isort = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-cov = "*"

[tool.poetry.scripts]
dev-health-ops = "dev_health_ops.cli:main"
dev-hops = "dev_health_ops.cli:main"

[project.scripts]
dev-health-ops = "dev_health_ops.cli:main"
dev-hops = "dev_health_ops.cli:main"
